portawaterperu: A Preliminary Review of Peruvian Potable Water System Data

Data for this project was obtained from @SIASAR, an information system containing data on rural water supply and sanitation services. Using SIASAR's "download data by country" tool, all available data for Peru (10 excel files) were downloaded. After examining the 10 excel files, only 5 pertained to potable water systems. Those 5 data sets were imported into R and subsequently empty values and unnecessary columns were deleted from them. Finally, the 5 data sets were combined into 1 data frame based on community ID. The combined, cleaned data set contains data from 32 communities.

Usage

portawaterperu

Format

A tibble with 6 rows and 18 variables

nombre

name of community water system

ID

identification number (unique to each water system)

pais

Country of data collection

divisiones

geographic division (region; province; district)

latitud

latitude of district

longitud

longitude of district

altitud

altitude of district

ano

year water system data was collected

comunidades

community/municipality/town name

PSE

water service provider

pob_servida

population number served by water service provider

viv_servidas

household number served by water service provider

tipo_gravedad

is the community served by a gravity distribution system?

tipo_bombeo

is the community served by a force/pump distribution system

tipo_pozo_manual

is the community served by a well system (communal collection point / no distribution pipes)

tipo_agua_lluvia

rain water (not sure in what regard this relates to water system won't use in analysis)

agua_epoca_seca

dry season water (not sure in what regard this relates to water system won't use in analysis)

agua_epoca_lluvia

rainy season water (not sure in what regard this relates to water system won't use in analysis)